Serialization: fix handling of numeric values and booleans
In Python 2, pylibmc essentially calls str() on the numeric value
and uses this byte string as what's stored in the cache. I didn't see
a very clean way to do this in Python 3, but this approach does work.

This does: numeric value -> text string containing ASCII representation
of the numeric value -> byte string that is the UTF-8 encoding of that
text. Boolean values are serialized the same way, just with a different
flag.