_PylibMC_map_str_keys: always return a valid dict
If something goes horribly wrong and we can't iterate over the keys
provided to one of the "multi" calls, we previously returned NULL for
the key translation mapping. This would then cause a segfault when
we try to use this object later.

Now, just return the valid (but empty) `dict`.