API changes to `get` and `__getitem__`
1. The `get` of _pylibmc.client now matches that of the dict object;
it takes a second argument, which is a default value to return if
the key is not found. If this argument is not passed, it defaults to
None, so this is fully backwards-compatible.

2. This default argument is used to implement `__getitem__` in a way
that unambiguously exposes key misses; previously, `__getitem__` could not
distinguish between a hit that returned a None and a miss. This is a (small)
API break; formerly, this:

client.set('key', None)
client['key']

would raise a KeyError, but now it returns None.